Use consistent metric for available memory
Using meminfo for available memory and MBean for total memory
means that we get inconsistent results when running in a container.
Meminfo will return the available memory of the machine, while MBean
will return the total memory of the container. This leads to Gradle
making bad memory-pressure related decisions.

Instead we now use MBean consistently and rely on the JVM getting
the right metrics for the environment we are in.